ci_requirements.txt
# dbt-coves==1.8.0
git+https://github.com/datacoves/dbt-coves.git@DCV-2515-dbt-coves-setup-datacoves
dbt-core==1.7.17
dbt-snowflake==1.7.5
pre-commit==3.7.1
apache-airflow==2.8.4